When rewriting the template page, cope with the Google Analytics snippet
The function that rewrote the links on the template page from the live
site was assuming that the <script> tag would always have a 'src'
attribute, but now that we've switched to the live site the template
includes the Google Analytics Javascript snippet which is just inline
rather than having a 'src' attribute. This commit changes the rewriting
so it doesn't fail if the attribute isn't present.