Quick fix to redirect to slug-based (not id-based) person URLs
The site is live now, but redirects back to the real site weren't
working because since this app was last updated the structure of person
URLs has changed. They're now based on slugs, not IDs.

This is a pretty quick and hacky change to accommodate that, because the
site's live now.